+++ src/core/evaluator.js
@@ -47,14 +47,7 @@ import {
   Ref,
   RefSet,
 } from "./primitives.js";
-import {
-  ErrorFont,
-  Font,
-  FontFlags,
-  getFontType,
-  IdentityToUnicodeMap,
-  ToUnicodeMap,
-} from "./fonts.js";
+import { ErrorFont, Font, FontFlags, getFontType } from "./fonts.js";
 import {
   getEncoding,
   MacRomanEncoding,
@@ -74,6 +67,7 @@ import {
   getSymbolsFonts,
 } from "./standard_fonts.js";
 import { getTilingPatternIR, Pattern } from "./pattern.js";
+import { IdentityToUnicodeMap, ToUnicodeMap } from "./to_unicode_map.js";
 import { isPDFFunction, PDFFunctionFactory } from "./function.js";
 import { Lexer, Parser } from "./parser.js";
 import {
